+++ roguelike/Vec2d.h
@@ -6,10 +6,10 @@
 class Vec2d {
 public:
 	Vec2d();
-	Vec2d(uint32_t x, uint32_t y);
+	Vec2d(size_t x, size_t y);
 	Vec2d(const Vec2d& other);
 
-	uint32_t x, y;
+	size_t x, y;
 
 	Vec2d operator +(const Vec2d& rhs);
 	Vec2d& operator +=(const Vec2d& rhs);
@@ -17,6 +17,6 @@ public:
 	Vec2d operator -(const Vec2d& rhs);
 	Vec2d& operator -=(const Vec2d& rhs);
 
-	uint32_t length();
-	static uint32_t distance(const Vec2d& lhs, const Vec2d& rhs);
+	size_t length();
+	static size_t distance(const Vec2d& lhs, const Vec2d& rhs);
 };
